RF-35: renaming thunks file
